Added classes and divs
I created the nav bar now, it might need some slight changes (for example moving the text closet to the symbol), otherwise i feel that it looks good.

I also started creating the item button that will contain the name of the beverages, the info, etc, which sooner will be connected to the database. This is not done yet, I couldn't figure out how to place the flag in the top right corner, so please help with that if you want to. Also, instead of margin-left, I want to use something that changes the button's place in relation to other objects, like the nav bar and the other item buttons, if you know what I mean.